Add integration test for property controller

+++ src/main/java/com/github/alina/repl/services/PropertyService.java
@@ -6,6 +6,6 @@ import com.github.alina.repl.models.entities.PropertyType;
 import java.util.List;
 
 public interface PropertyService {
-    List<PropertyDTO> getProperties(PropertyType propertyType, String city, Long price, int bedrooms);
+    List<PropertyDTO> getProperties(PropertyType propertyType, String city, Long price, Long bedrooms);
 
 }
